requirements.txt
binance-connector
python-telegram-bot
schedule~=1.0.0